requirements.txt
django-args==1.4.0
certifi
chardet
clint
coverage==4.0.3
coveralls==1.1
curdling==0.4.0
distlib==0.1.2
docopt==0.6.2
docutils==0.12
funcsigs==0.4
httpretty==0.8.14
idna==2.7
m2r==0.2.1
mysql_client
django-allauth
django-adminlte-3
mistune==0.8.4
mock==1.3.0
pbr==1.8.1
pkginfo==1.4.2
requests==2.20.0
requests-toolbelt==0.8.0
simplejson==3.8.2
six==1.10.0
django-excel
sure==1.2.24
django-tinymce
django-widget-tweaks
django_static_fontawesome
django-qr_code
tqdm==4.25.0
django_countries
paystackapi
twine==1.11.0
django_hosts
urllib3==1.24.2
django-qr-code
django-wkhtmltopdf
django-mathfilters
django-imagekit
pillow
django-cuser
whitenoise
wkhtmltopdf